+++ src/action/target.js
@@ -30,6 +30,11 @@ const actions = createActions({
     * @param {object} options = { id, target, selector, editing }
     * @returns {object}
     */
+  SET_TARGET: ( options ) => validate( options, { ...I.ENTITY, ...I.TARGET, ...I.UPDATE }),
+  /**
+    * @param {object} options = { id, target, selector, editing }
+    * @returns {object}
+    */
   UPDATE_TARGET: ( options ) => validate( options, { ...I.ENTITY, ...I.TARGET, ...I.UPDATE }),
   /**
     * @param {object} ref = { id }